refactor: consolidate Deck model in schema
