+++ src/entities/deck/@x/card.ts
@@ -1,3 +1,3 @@
-import type { Deck } from "../model/deck";
+import type { Deck } from "../model/schema";
 
 export type DeckForCard = Pick<Deck, "id" | "uid">;
