+++ src/entities/deck/index.ts
@@ -1,8 +1,7 @@
 export { CATEGORY, getCategory, isHighlightLanguage } from "./model/category";
 export type { Category } from "./model/category";
 export { generateDeckId } from "./api/firestore";
-export { createDeck } from "./model/deck";
-export { createDeckSchema, deleteDeckSchema, editDeckSchema } from "./model/schema";
+export { createDeck, createDeckSchema, deleteDeckSchema, editDeckSchema } from "./model/schema";
 export type {
   CreateDeckInput,
   Deck,
