+++ src/entities/deck/model/schema.spec.ts
@@ -2,7 +2,27 @@ import { describe, expect, it } from "vitest";
 
 import { createDeck as createDeckFixture } from "@/test/factories";
 
-import { createDeckSchema, deleteDeckSchema, editDeckSchema } from "./schema";
+import { createDeck, createDeckSchema, deleteDeckSchema, editDeckSchema } from "./schema";
+
+describe("createDeck", () => {
+  it("creates a deck with entity defaults and an injected id", () => {
+    expect(createDeck({ name: "name" }, "uid", () => "deck-id")).toEqual({
+      name: "name",
+      id: "deck-id",
+      uid: "uid",
+      createdAt: 0,
+      updatedAt: 0,
+      deletedAt: null,
+      scoreMax: null,
+      scoreMin: null,
+      isPublic: false,
+      selectedTags: [],
+      tagAndFilter: false,
+      convertToBr: false,
+      category: "",
+    });
+  });
+});
 
 describe("Deck operation schemas", () => {
   const deck = createDeckFixture({ id: "deck", uid: "uid-a" });
